fix: rabbit rants
