+++ src/ui/admin_state.rs
@@ -1,26 +1,5 @@
+use crate::shared::permissions::SolverPermission;
 use crate::ui::KeyInputState;
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum SolverPermission {
-    Read,
-    ReadWrite,
-}
-
-impl SolverPermission {
-    pub const fn toggle(self) -> Self {
-        match self {
-            Self::Read => Self::ReadWrite,
-            Self::ReadWrite => Self::Read,
-        }
-    }
-
-    pub const fn as_label(self) -> &'static str {
-        match self {
-            Self::Read => "read",
-            Self::ReadWrite => "read-write",
-        }
-    }
-}
 
 #[derive(Clone, Debug)]
 pub struct AddSolverState {
