+++ src/ui/state.rs
@@ -1,4 +1,5 @@
-pub use crate::ui::admin_state::{AddSolverState, SolverPermission};
+pub use crate::shared::permissions::SolverPermission;
+pub use crate::ui::admin_state::AddSolverState;
 pub use crate::ui::app_state::{AppState, UiMode};
 pub use crate::ui::chat::{
     AdminChatLastSeen, AdminChatUpdate, ChatAttachment, ChatAttachmentType, ChatParty, ChatSender,
